+++ .gitmodules
@@ -1,6 +1,6 @@
 [submodule "luajit-2.0"]
-	path = luajit-2.0
+	path = app/src/main/luajit-2.0
 	url = https://github.com/LuaJIT/LuaJIT
 [submodule "jni/7z"]
-	path = jni/7z
+	path = app/src/main/jni/7z
 	url = https://github.com/pornel/7z.git
