--- AndroidManifest.xml
@@ -1,91 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-      package="org.koreader.launcher"
-      android:versionCode="4"
-      android:versionName="1.4">
-    <uses-sdk android:minSdkVersion="9" />
-    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
-    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
-    <uses-permission android:name="android.permission.WRITE_SETTINGS" />
-    <uses-permission android:name="android.permission.INTERNET" />
-    <uses-permission android:name="android.permission.ACCESS_WIFI_STATE"/>
-    <uses-permission android:name="android.permission.CHANGE_WIFI_STATE"/>
-    <application android:icon="@drawable/icon" android:label="@string/app_name"
-        android:hasCode="true" android:debuggable="true"
-        android:largeHeap="true" android:vmSafeMode="true">
-        <activity android:name="org.koreader.launcher.MainActivity"
-                android:label="@string/app_name"
-                android:screenOrientation="portrait"
-                android:launchMode="singleInstance"
-                android:theme="@style/Fullscreen">
-            <meta-data android:name="android.app.lib_name"
-                    android:value="luajit" />
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-            <intent-filter>
-                <action android:name="android.intent.action.VIEW" />
-                <category android:name="android.intent.category.DEFAULT" />
-                <data android:mimeType="application/pdf" />
-                <data android:mimeType="image/djvu" />
-                <data android:mimeType="image/vnd.djvu" />
-                <data android:mimeType="image/x-djvu" />
-                <data android:mimeType="application/x-cbz" />
-                <data android:mimeType="application/epub+zip" />
-                <data android:mimeType="application/x-fb2" />
-                <data android:mimeType="application/x-mobipocket-ebook" />
-                <data android:mimeType="application/msword" />
-                <data android:mimeType="application/x-chm" />
-                <data android:mimeType="text/plain" />
-            </intent-filter>
-            <intent-filter>
-                <action android:name="android.intent.action.VIEW" />
-                <category android:name="android.intent.category.DEFAULT" />
-                <data android:host="*" />
-                <data android:scheme="file" />
-                <data android:mimeType="*/*" />
-                <!-- the second and third lines for each respective extension are a workaround for file patterns with dots in them, like /my.folder/test.pdf or  "Joe W. Writer-Love in Somewhere.pdf" -->
-                <data android:pathPattern=".*\\.pdf" />
-                <data android:pathPattern=".*\\..*\\.pdf" />
-                <data android:pathPattern=".*\\..*\\..*\\.pdf" />
-                <data android:pathPattern=".*\\.djvu" />
-                <data android:pathPattern=".*\\..*\\.djvu" />
-                <data android:pathPattern=".*\\..*\\..*\\.djvu" />
-                <data android:pathPattern=".*\\.epub" />
-                <data android:pathPattern=".*\\..*\\.epub" />
-                <data android:pathPattern=".*\\..*\\..*\\.epub" />
-                <data android:pathPattern=".*\\.cbz" />
-                <data android:pathPattern=".*\\..*\\.cbz" />
-                <data android:pathPattern=".*\\..*\\..*\\.cbz" />
-                <data android:pathPattern=".*\\.zip" />
-                <data android:pathPattern=".*\\..*\\.zip" />
-                <data android:pathPattern=".*\\..*\\..*\\.zip" />
-                <data android:pathPattern=".*\\.xps" />
-                <data android:pathPattern=".*\\..*\\.xps" />
-                <data android:pathPattern=".*\\..*\\..*\\.xps" />
-                <data android:pathPattern=".*\\.fb2" />
-                <data android:pathPattern=".*\\..*\\.fb2" />
-                <data android:pathPattern=".*\\..*\\..*\\.fb2" />
-                <data android:pathPattern=".*\\.mobi" />
-                <data android:pathPattern=".*\\..*\\.mobi" />
-                <data android:pathPattern=".*\\..*\\..*\\.mobi" />
-                <data android:pathPattern=".*\\.chm" />
-                <data android:pathPattern=".*\\..*\\.chm" />
-                <data android:pathPattern=".*\\..*\\..*\\.chm" />
-                <data android:pathPattern=".*\\.pdb" />
-                <data android:pathPattern=".*\\..*\\.pdb" />
-                <data android:pathPattern=".*\\..*\\..*\\.pdb" />
-                <data android:pathPattern=".*\\.doc" />
-                <data android:pathPattern=".*\\..*\\.doc" />
-                <data android:pathPattern=".*\\..*\\..*\\.doc" />
-                <data android:pathPattern=".*\\.tcr" />
-                <data android:pathPattern=".*\\..*\\.tcr" />
-                <data android:pathPattern=".*\\..*\\..*\\.tcr" />
-                <data android:pathPattern=".*\\.rtf" />
-                <data android:pathPattern=".*\\..*\\.rtf" />
-                <data android:pathPattern=".*\\..*\\..*\\.rtf" />
-            </intent-filter>
-        </activity>
-    </application>
-</manifest>
