+++ app/src/main/java/org/koreader/launcher/MainActivity.java
@@ -4,22 +4,22 @@ import android.app.NativeActivity;
 import android.content.ClipData;
 import android.content.ClipDescription;
 import android.content.ClipboardManager;
-import android.provider.Settings;
-import android.view.WindowManager;
-import android.os.BatteryManager;
-import android.content.IntentFilter;
-import android.content.Intent;
 import android.content.Context;
+import android.content.Intent;
+import android.content.IntentFilter;
+import android.graphics.Point;
+import android.graphics.Rect;
+import android.net.wifi.WifiManager;
+import android.os.BatteryManager;
 import android.os.Bundle;
 import android.os.Handler;
-import android.view.View;
-import android.util.Log;
-import android.graphics.Point;
-import android.view.Display;
-import android.graphics.Rect;
-import android.view.Window;
+import android.provider.Settings;
 import android.util.DisplayMetrics;
-import android.net.wifi.WifiManager;
+import android.util.Log;
+import android.view.Display;
+import android.view.View;
+import android.view.Window;
+import android.view.WindowManager;
 
 import java.util.concurrent.CountDownLatch;
 
@@ -27,11 +27,12 @@ public class MainActivity extends NativeActivity {
 
     private final static int SDK_INT = android.os.Build.VERSION.SDK_INT;
     private final static String LOGGER_NAME = "luajit-launcher";
-    private FramelessProgressDialog dialog;
 
     static {
         System.loadLibrary("luajit");
     }
+
+    private FramelessProgressDialog dialog;
 
     public MainActivity() {
         super();
@@ -48,7 +49,12 @@ public class MainActivity extends NativeActivity {
         super.onResume();
 
         final Handler handler = new Handler();
-        handler.postDelayed(() -> setFullscreenLayout(), 500);
+        handler.postDelayed(new Runnable() {
+            @Override
+            public void run() {
+                setFullscreenLayout();
+            }
+        }, 500);
     }
 
     private void setFullscreenLayout() {
@@ -56,47 +62,34 @@ public class MainActivity extends NativeActivity {
             getWindow().getDecorView().setSystemUiVisibility(View.STATUS_BAR_HIDDEN);
         } else if (SDK_INT >= 16 && SDK_INT < 19) {
             getWindow().getDecorView().setSystemUiVisibility(
-                    View.SYSTEM_UI_FLAG_FULLSCREEN | View.SYSTEM_UI_FLAG_LOW_PROFILE);
+                View.SYSTEM_UI_FLAG_FULLSCREEN | View.SYSTEM_UI_FLAG_LOW_PROFILE);
         } else if (SDK_INT >= 19) {
             getWindow().getDecorView().setSystemUiVisibility(
-                    View.SYSTEM_UI_FLAG_LAYOUT_STABLE |
-                            View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION |
-                            View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN |
-                            View.SYSTEM_UI_FLAG_HIDE_NAVIGATION |
-                            View.SYSTEM_UI_FLAG_FULLSCREEN |
-                            View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY);
+                View.SYSTEM_UI_FLAG_LAYOUT_STABLE |
+                    View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION |
+                    View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN |
+                    View.SYSTEM_UI_FLAG_HIDE_NAVIGATION |
+                    View.SYSTEM_UI_FLAG_FULLSCREEN |
+                    View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY);
         }
-    }
-
-    public void setScreenBrightness(final int brightness) {
-        runOnUiThread(() -> {
-            try {
-                //this will set the manual mode (set the automatic mode off)
-                Settings.System.putInt(getContentResolver(),
-                        Settings.System.SCREEN_BRIGHTNESS_MODE,
-                        Settings.System.SCREEN_BRIGHTNESS_MODE_MANUAL);
-
-                Settings.System.putInt(getContentResolver(),
-                        Settings.System.SCREEN_BRIGHTNESS, brightness);
-            } catch (Exception e) {
-                Log.v(LOGGER_NAME, e.toString());
-            }
-        });
     }
 
     public int getScreenBrightness() {
         final Box<Integer> result = new Box<Integer>();
         final CountDownLatch cd = new CountDownLatch(1);
-        runOnUiThread(() -> {
-            try {
-                result.value = new Integer(Settings.System.getInt(
+        runOnUiThread(new Runnable() {
+            @Override
+            public void run() {
+                try {
+                    result.value = new Integer(Settings.System.getInt(
                         getContentResolver(),
                         Settings.System.SCREEN_BRIGHTNESS));
-            } catch (Exception e) {
-                Log.v(LOGGER_NAME, e.toString());
-                result.value = new Integer(0);
+                } catch (Exception e) {
+                    Log.v(LOGGER_NAME, e.toString());
+                    result.value = new Integer(0);
+                }
+                cd.countDown();
             }
-            cd.countDown();
         });
         try {
             cd.await();
@@ -111,9 +104,28 @@ public class MainActivity extends NativeActivity {
         return result.value.intValue();
     }
 
+    public void setScreenBrightness(final int brightness) {
+        runOnUiThread(new Runnable() {
+            @Override
+            public void run() {
+                try {
+                    //this will set the manual mode (set the automatic mode off)
+                    Settings.System.putInt(getContentResolver(),
+                        Settings.System.SCREEN_BRIGHTNESS_MODE,
+                        Settings.System.SCREEN_BRIGHTNESS_MODE_MANUAL);
+
+                    Settings.System.putInt(getContentResolver(),
+                        Settings.System.SCREEN_BRIGHTNESS, brightness);
+                } catch (Exception e) {
+                    Log.v(LOGGER_NAME, e.toString());
+                }
+            }
+        });
+    }
+
     public int getBatteryLevel() {
         Intent intent = this.registerReceiver(null,
-                new IntentFilter(Intent.ACTION_BATTERY_CHANGED));
+            new IntentFilter(Intent.ACTION_BATTERY_CHANGED));
         int level = intent.getIntExtra(BatteryManager.EXTRA_LEVEL, 0);
         int scale = intent.getIntExtra(BatteryManager.EXTRA_SCALE, 100);
         return (level * 100) / scale;
@@ -121,21 +133,30 @@ public class MainActivity extends NativeActivity {
 
     public int isCharging() {
         Intent intent = this.registerReceiver(null,
-                new IntentFilter(Intent.ACTION_BATTERY_CHANGED));
+            new IntentFilter(Intent.ACTION_BATTERY_CHANGED));
         int plugged = intent.getIntExtra(BatteryManager.EXTRA_PLUGGED, -1);
         return plugged == BatteryManager.BATTERY_PLUGGED_AC
-                || plugged == BatteryManager.BATTERY_PLUGGED_USB ? 1 : 0;
+            || plugged == BatteryManager.BATTERY_PLUGGED_USB ? 1 : 0;
     }
 
     public void showProgress(final String title, final String message) {
-        runOnUiThread(() -> {
-            dialog = FramelessProgressDialog.show(MainActivity.this, title, message, true, false);
+        runOnUiThread(new Runnable() {
+            @Override
+            public void run() {
+                dialog = FramelessProgressDialog.show(MainActivity.this,
+                    title, message, true, false);
+            }
         });
     }
 
     public void dismissProgress() {
-        runOnUiThread(() -> {
-            if (dialog != null && dialog.isShowing()) dialog.dismiss();
+        runOnUiThread(new Runnable() {
+            @Override
+            public void run() {
+                if (dialog != null && dialog.isShowing()) {
+                    dialog.dismiss();
+                }
+            }
         });
     }
 
@@ -145,30 +166,38 @@ public class MainActivity extends NativeActivity {
 
 
     public void setWifiEnabled(final boolean enabled) {
-      runOnUiThread(() -> this.getWifiManager().setWifiEnabled(enabled));
+        runOnUiThread(new Runnable() {
+            @Override
+            public void run() {
+                getWifiManager().setWifiEnabled(enabled);
+            }
+        });
     }
 
     public int isWifiEnabled() {
-        return this.getWifiManager().isWifiEnabled() ? 1 : 0;
+        return getWifiManager().isWifiEnabled() ? 1 : 0;
     }
 
     private WifiManager getWifiManager() {
-        return (WifiManager) getSystemService(WIFI_SERVICE);
+        return (WifiManager) this.getSystemService(Context.WIFI_SERVICE);
     }
 
     public void setKeepScreenOn(final boolean keepOn) {
         final CountDownLatch cd = new CountDownLatch(1);
-        runOnUiThread(() -> {
-            try {
-                if (keepOn) {
-                    getWindow().addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
-                } else {
-                    getWindow().clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
+        runOnUiThread(new Runnable() {
+            @Override
+            public void run() {
+                try {
+                    if (keepOn) {
+                        getWindow().addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
+                    } else {
+                        getWindow().clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
+                    }
+                } catch (Exception e) {
+                    Log.v(LOGGER_NAME, e.toString());
                 }
-            } catch (Exception e) {
-                Log.v(LOGGER_NAME, e.toString());
+                cd.countDown();
             }
-            cd.countDown();
         });
         try {
             cd.await();
@@ -179,38 +208,27 @@ public class MainActivity extends NativeActivity {
 
     public void setFullscreen(final boolean fullscreen) {
         final CountDownLatch cd = new CountDownLatch(1);
-        runOnUiThread(() -> {
-            try {
-                WindowManager.LayoutParams attrs = getWindow().getAttributes();
-                if (fullscreen) {
-                    attrs.flags |= WindowManager.LayoutParams.FLAG_FULLSCREEN;
-                } else {
-                    attrs.flags &= ~WindowManager.LayoutParams.FLAG_FULLSCREEN;
+        runOnUiThread(new Runnable() {
+            @Override
+            public void run() {
+                try {
+                    WindowManager.LayoutParams attrs = getWindow().getAttributes();
+                    if (fullscreen) {
+                        attrs.flags |= WindowManager.LayoutParams.FLAG_FULLSCREEN;
+                    } else {
+                        attrs.flags &= ~WindowManager.LayoutParams.FLAG_FULLSCREEN;
+                    }
+                    getWindow().setAttributes(attrs);
+                } catch (Exception e) {
+                    Log.v(LOGGER_NAME, e.toString());
                 }
-                getWindow().setAttributes(attrs);
-            } catch (Exception e) {
-                Log.v(LOGGER_NAME, e.toString());
+                cd.countDown();
             }
-            cd.countDown();
-
         });
         try {
             cd.await();
         } catch (InterruptedException ex) {
         }
-    }
-
-    public void setClipboardText(String text) {
-        runOnUiThread(() -> {
-            if (SDK_INT < android.os.Build.VERSION_CODES.HONEYCOMB) {
-                android.text.ClipboardManager clipboard = (android.text.ClipboardManager) getSystemService(CLIPBOARD_SERVICE);
-                clipboard.setText(text);
-            } else {
-                ClipboardManager clipboard = (ClipboardManager) getSystemService(CLIPBOARD_SERVICE);
-                ClipData clip = ClipData.newPlainText("Koreader_clipboard", text);
-                clipboard.setPrimaryClip(clip);
-            }
-        });
     }
 
     public String getClipboardText() {
@@ -225,6 +243,26 @@ public class MainActivity extends NativeActivity {
             }
         }
         return "";
+    }
+
+    public void setClipboardText(String text) {
+        runOnUiThread(new Runnable() {
+            @Override
+            public void run() {
+                try {
+                    if (SDK_INT < android.os.Build.VERSION_CODES.HONEYCOMB) {
+                        android.text.ClipboardManager clipboard = (android.text.ClipboardManager) getSystemService(Context.CLIPBOARD_SERVICE);
+                        clipboard.setText(text);
+                    } else {
+                        ClipboardManager clipboard = (ClipboardManager) getSystemService(Context.CLIPBOARD_SERVICE);
+                        ClipData clip = ClipData.newPlainText("KOReader_clipboard", text);
+                        clipboard.setPrimaryClip(clip);
+                    }
+                } catch (Exception e) {
+                    Log.v(LOGGER_NAME, e.toString());
+                }
+            }
+        });
     }
 
     public boolean hasClipboardText() {
